+++ dump_things_service/tests/test_extract_inline.py
@@ -159,6 +159,7 @@ def test_inline_extraction_locally(dump_stores_simple):
         model=MockedModule(),
         pid_mapping_function=None,
         suffix='yaml',
+        sort_keys=[],
     )
     records = store.extract_inlined(inlined_object, 'hans')
     _check_result_objects(records, tree)
@@ -190,6 +191,7 @@ def test_dont_extract_empty_things_locally(dump_stores_simple):
         model=MockedModule(),
         pid_mapping_function=None,
         suffix='yaml',
+        sort_keys=[],
     )
     records = store.extract_inlined(empty_inlined_object, 'dieter')
     assert len(records) == 1